create group: Extract user selection into UserPickerCard
Make the user picking functionality into a new reusable component.
`GroupCard` becomes the new `UserPickerCard` but `GroupScreen`
retains the code that narrows to a new group.
