+++ src/user-picker/AvatarItem.js
