+++ src/user-picker/AvatarList.js
